python/gpu-gen/requirements.txt
scipy>=1.3.3
requests>=2.22.0
Pillow>=6.2.1
h5py>=2.9.0
imageio>=2.9.0
imageio-ffmpeg>=0.4.2
tqdm>=4.49.0
google-cloud-pubsub>=2.8.0
python-dotenv
psycopg2
wand~=0.6.7
google-cloud-firestore>=2.3.4
google-cloud-storage